sched/fair: Let new tasks go through the energy-aware wakeup path
When the energy-aware wakeup path is enabled in a lightly loaded scenario,
the most energy-inefficient (and usually most capable) CPUs are likely to
remain mostly idle. If the placement decision of new fork tasks if left to
find_idlest_cpu(), the least loaded CPU will be selected, which has a high
chance of being one of the aforementioned energy-inefficient CPUs. As a
consequence, the initial util_avg of the task will be set to half of the
spare capacity of this CPU and might results in a very power-expensive
frequency adjustment if schedutil is being used.

As a workaround, this commit lets all tasks go through the energy-aware
wakeup path, including new tasks which don't have a representative
util_avg yet. With a null util_avg, find_energy_efficient_cpu() will
return prev_cpu as the impact on energy of the new task cannot be
foreseen. While this can clearly have a negative impact on performance,
the main use-case targeted here is the one of very low CPU demand where
the probability of spawning heavy tasks is generally low as well. If
this assumption happens to be incorrect, or in the case of more CPU
intensive workloads, the overutilized threshold should be crossed
promptly, hence giving the load balance and the usual wakeup path the
opportunity to balance things out.

Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>