ANDROID: update cpu capacities after sched domains are built
If we're only left with one big CPU in a cluster (either there was only
one to begin with or the others have been hotplugged out), the MC level
seen from the perspective of that CPU, will only contain one group (that
of the CPU) and the LOAD_BALANCE flag will be cleared at that level.

That will result in update_cpu_capacity never being called for that big
CPU and its capacity will never be updated. Also max_cpu_capacity will
never factor in the existance of this CPU and will point to one of the
LITTLE CPUs and its capacity as the maximum capacity in the system.

Although init_sched_groups_capacity should handle this in build_sched_domains,
it does that while still having the CPU RQs poining to a default root
domain. This one will be updated instead of the root domain that will
later be attached to the RQs.

Move the initialisation of capacities (by calling update_group_capacity)
after the final root domain is attached to RQs.
Set the min and max original capacity values after update_cpu_capacity
was called for each CPU and cpu_capacity_orig was updated.

Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>