ANDROID: update_sched_groups_capacity when switched_to_fair
RT pressure contributes to a CPU's capacity as it can be seen in
update_cpu_capacity, but this function is not called often enough to
quickly reflect updates in RT pressure.

Update CPU and group capacities when switching from (possibly) an RT
tasks to a fair one to improve reactivity.
This in only meant for experimentation currently and it requires
more investigation.

Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>