sched: reduce the periodic update duration
Instead of using the cfs_rq_is_decayed() which monitors all *_avg
and *_sum, we create a cfs_rq_has_blocked() which only takes care of
util_avg and load_avg. We are only interested by these 2 values which are
decaying faster than the *_sum so we can stop the periodic update earlier.

Cc: valentin.schneider@arm.com
Cc: mingo@kernel.org
Cc: dietmar.eggemann@arm.com
Cc: brendan.jackman@arm.com
Cc: morten.rasmussen@foss.arm.com
Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: http://lkml.kernel.org/r/1518422874-13216-3-git-send-email-vincent.guittot@linaro.org