sched: Rework root_domain->overload flagging
The rq->rd->overload flag is currently:
- raised when going through add_nr_running() and rq->nr_running
becomes => 2
- lowered when going through update_sd_lb_stats() and no group had
rq->nr_running > 1

If raised, NEWLY_IDLE load balance (IOW idle balance) will be enabled.

However, if we want to enable idle balance for other reasons, we'd have
to set the 'overload' flag elsewhere and it wouldn't make sense to call it
"overload" anymore. Furthermore, we wouldn't be able to determine
why it was raised in the first place.

This commit replaces the 'overload' bool with a 'balance_status' int in
which specific flags can be stored. For now, this is only LBS_OVERLOAD.

Signed-off-by: Valentin Schneider <valentin.schneider@arm.com>