TEST: faster CPU selection
DO NOT SEND

Inside a cluster (=freq domain), the CPU with the max spare cap is
_always_ going to be in the set of best CPUs. What I mean by "set of
best CPUs" is the set of CPUs candidates for which the system energy is
the same (and the lowest). This is a correct assumption with the
simplified EM, not with the EM in product

Finding the CPU with max spare cap can be done quickly so let's do an
energy calculation only for this one, for each cluster. That means only
3 (prev_cpu + little cluster + big cluster) calls to compute_energy()
on big.LITTLE.

If that's all good, this patch should be squashed into "sched/fair: Select
an energy-efficient CPU on task wake-up"