sched/fair: maximise energy efficiency for !prefer_idle tasks
find_best_target is currently split into code handling latency sensitive
tasks and code handling non-latency sensitive tasks based on the value
of the prefer_idle flag.
Another differentiation was done for boosted tasks, preferring to start
with higher-capacity CPU when boosted, and with more efficient CPUs when
not boosted. This additional differentiation could result in the
selection of a big CPU for boosted non-prefer idle tasks and the selection
of a little CPU for prefer_idle but non-boosted tasks.

It makes more sense for non-latency sensitive tasks to always try to
minimise capacity_orig of selected CPUs (maximise energy efficiency).

There is no reason for boosted and !prefer_idle cases to be given the
advantage of potentially selecting a big CPU, which is the advantage
they would obtain by ordering them from big to LITTLE for boosted tasks.
The other mechanisms in place set for boosted tasks (making sure we
select a CPU that fits the task) should be enough. Also, by choosing a
CPU with maximum spare capacity we also cover the preference towards
spreading tasks, rather than packing them, which actually improves the
chances for a boostd task to get better performance due to potential
reduced preemption.

Taking in part each of the two cases:
 1. Selection of a backup idle CPU - Non prefer_idle tasks should prefer
    more energy efficient CPUs when there are idle CPUs in the system,
    independent of the order given by the presence of a booste margin.
    This was the behavior for !sysctl_sched_cstate_aware and I believe
    should be the behaviour for sysctl_sched_cstate_aware set as well,
    given that we should prefer a more efficient CPU even if it's in a
    deeper idle state.

 2. Selection of an active target CPU: There is no reason for boosted
    tasks to benefit from a higher chance to big CPUs with higher spare
    which is provided by ordering CPUs from bigs to littles. The goal
    here should be to maximise energy efficiency (minimise capacity) as
    well.

Change-Id: I3b97010e682674420015e771f0717192444a63a2
Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>
Reported-by: Viresh Kumar <viresh.kumar@linaro.org>
Reported-by: Leo Yan <leo.yan@linaro.org>