sched/core: extend sched_setattr to support util_clamp
The SCHED_DEADLINE scheduling class provides an advanced and formal
model to define tasks requirements which can be translated into proper
decisions for both task placements and frequencies selections.
Other classes have a more simplified model which is essentially based on
the relatively simple concept of POSIX priorities.

This simple model however does not allow to better exploit some of the
most advanced features of the Linux scheduler, like for example driving
frequencies selection via the schedutil cpufreq governor.
However, also for non SCHED_DEADLINE tasks it's still interesting to
define tasks properties which can be used to better support certain
scheduler decisions.

Utilization clamping already provide such a support on systems where
CGroups are in use. However, utilization clamping can also be a per-task
attribute which can be used also when CGroups are not in use.

Let's introduce a new API to set util_clamp values for a specified task
by extending sched_setattr, a syscall which already allows to define
task specific properties to support different scheduling classes.
For non SCHED_DEADLINE tasks, a new pair of attributes allows to specify
the minimum and maximum utilization which the scheduler should consider
for a task.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Tejun Heo <tj@kernel.org>
Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: Vincent Guittot <vincent.guittot@linaro.org>
Cc: Viresh Kumar <viresh.kumar@linaro.org>
Cc: Paul Turner <pjt@google.com>
Cc: Joel Fernandes <joelaf@google.com>
Cc: Steve Muckle <smuckle@google.com>
Cc: Juri Lelli <juri.lelli@redhat.com>
Cc: Dietmar Eggemann <dietmar.eggemann@arm.com>
Cc: Morten Rasmussen <morten.rasmussen@arm.com>
Cc: linux-kernel@vger.kernel.org
Cc: linux-pm@vger.kernel.org

---
This patch introduces only the sched_setattr extension to make it easy
reviewing this proposal without being distracted from the details of how
we use the new task attributes. The actual usage is introduced by the
following patch.

The solution proposed here expose the concept of "utilization" to
userspace. This should be a quite generic concept, maybe if we abstract
it a bit to be considered just as a percentage of the CPU capacity, and
thus in the range [0..100] instead of [0..SCHED_CAPACITY_SCALE] as it is
now.

If such a defined utilization should still be considered too much an
implementation detail, a possible alternative proposal can be that do
"recycle" the usage of:
   sched_runtime
   sched_period
to be translated internally into a proper utilization.
Such a model would allow to have a more abstract description of the
expected task utilization and, combined with in-kernel knowledge
such as the PELT half-life, can probably be translated into a better
min/max utilization clamp value.

For this first proposal, I've opted to present the most simple solution
based on a new "abstract" utilization metric.