sched/core: aggregate TG and Task's util_clamp values
When a task's util_clamp value is configured via sched_setattr, this
value has to be properly accounted in the corresponding clamp group
every time the task is enqueue and dequeued. When cgroups are also in
use, per-task clamp values have to be aggregated to those of the cpu's
controller CGroup in which the task is currently living.

Let's first ensure that, whenever a task is assigned a specific
clamp_value, this is properly translated into a unique clamp group to be
used in the fast-path (i.e. at enqueue/dequeue time). This is done by
slightly refactoring uclamp_group_get to accept a *task_struct alongside
a *cgroup_subsys_state.

When uclamp_group_get is called with a valid *task_struct, a clamp group
is assigned to the task, which is possibly different than the cgroup
clamp group where the task lives. If the task is currently RUNNABLE, the
accounting is also properly updated via uclamp_task_update_active().

Finally, uclamp_cpu_get() is update to provide an aggregation between
the task and the TG clamp values. Every time a task is enqueued, it will
be accounted in the clamp_group which defines the smaller clamp value
between the task and the TG's ones.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Tejun Heo <tj@kernel.org>
Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: Vincent Guittot <vincent.guittot@linaro.org>
Cc: Viresh Kumar <viresh.kumar@linaro.org>
Cc: Paul Turner <pjt@google.com>
Cc: Joel Fernandes <joelaf@google.com>
Cc: Steve Muckle <smuckle@google.com>
Cc: Juri Lelli <juri.lelli@redhat.com>
Cc: Dietmar Eggemann <dietmar.eggemann@arm.com>
Cc: Morten Rasmussen <morten.rasmussen@arm.com>
Cc: linux-kernel@vger.kernel.org
Cc: linux-pm@vger.kernel.org