sched/core: uclamp: use TG's clamps to restrict Task's clamps
When a task's util_clamp value is configured via sched_setattr(2), this
value has to be properly accounted in the corresponding clamp group
every time the task is enqueue and dequeued. When cgroups are also in
use, per-task clamp values have to be aggregated to those of the CPU's
controller's Task Group (TG) in which the task is currently living.

Let's update uclamp_cpu_get() to provide an aggregation between the task
and the TG clamp values. Every time a task is enqueued, it will be
accounted in the clamp_group which defines the smaller clamp between the
task specific value one and its TG value.

This also mimics what already happen for a task's CPU affinity mask when
the task is also living in a cpuset.  The overall idea is that: CGroups
attributes are always used to restrict the per-task attributes.
Thus, this implementation allow to:
1. ensure CGroups clamps are always used to restrict task specific
   requests, i.e. boosted only up to a granted value or clamped at least
   to a certain value
2. implements a nice policy, where tasks are still allowed to request
   less then whan enforced by their current TG

For this mecanism to work properly, we need to implement a concept of
"effective" clamp group, which is used to track the currently most
restrictive clamp value each task is subject to.
The effective clamp is computed at enqueue time, by using an additional
   task_struct::uclamp_group_id
to keep track of the clamp group in which each task is currently
accounted into. This solution allow to update tasks constrains on
demand, only when they became RUNNABLE, to always get the less
restrictive clamp depending on the current TG's settings.

This solution allows also to better decouple the slow-path, where task
specific and task group clamp values are updated, from the fast-path,
where the most appropriate clamp value is tracked by refcounting clamp
groups.

For consistency purposes, as well as to properly inform userspace, the
sched_getattr(2) call is updated to always return the properly
aggregated constrains as described above. This will also make
sched_getattr(2) a convenient userpace API to know the utilization
constraints enforced on a task by the CGroups's CPU controller.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Tejun Heo <tj@kernel.org>
Cc: Paul Turner <pjt@google.com>
Cc: Todd Kjos <tkjos@google.com>
Cc: Joel Fernandes <joelaf@google.com>
Cc: Steve Muckle <smuckle@google.com>
Cc: Juri Lelli <juri.lelli@redhat.com>
Cc: Dietmar Eggemann <dietmar.eggemann@arm.com>
Cc: Morten Rasmussen <morten.rasmussen@arm.com>
Cc: linux-kernel@vger.kernel.org
Cc: linux-pm@vger.kernel.org

---
Changes in v2:

 OSPM discussion:
 - implement a "nice" semantics where cgroup clamp values are always
   used to restrict task specific clamp values, i.e. tasks running on a
   TG are only allowed to demote themself.

 Other:
 - rabased on v4.18-rc4
 - this code has been split from a previous patch to simplify the review