sched/fair: remove order from CPU selection
find_best_target is currently split into code handling latency sensitive
tasks and code handling non-latency sensitive tasks based on the value
of the prefer_idle flag.
Another differentiation is done for boosted tasks, preferring to start
with higher-capacity CPU when boosted, and with more efficient CPUs when
not boosted. This additional differentiation is obtained by imposing an
order when considering CPUs for selection. This order is determined in
typical big.LITTLE systems by the start point (the CPU with the maximum
or minimum capacity) and by the order of big and little CPU groups
provided in the sched domain hierarchy.

But it's not always guaranteed that the sched domain hierachy will give
us a sorted list of CPU groups based on their maximum capacities when
dealing with multiple cluster systems (>2).
For example, if we consider a system with three groups of CPUs (LITTLEs,
mediums, bigs), the sched domain hierarchy might provide the following
order of scheduling groups for CPU selection in find_best_target, for a
prefer_idle and boosted task:
   big CPUs -> LITTLE CPUs -> medium CPUs.
If the big CPUs are not idle, but there are a few LITTLEs and mediums
as idle CPUs, by returning the first idle CPU, we will be incorrectly
prefering a lower capacity CPU over a higher capacity CPU.

In order to eliminate this reliance on considering CPUs in order,
based on their capacities, this solution does the following:
1. Iterates though all candidate CPUs for all cases.
2. Minimises or maximises the capacity of the considered CPU, depending
on prefer_idle and boost information.

Taking in part each of the four possible cases we analyse the
implementation and impact of this solution:

1. prefer_idle and boosted
This type of tasks needs to favour the selection of a reserved CPU and
therefore we maintain as start point the use of biggest CPU in the system.
But we iterate though all CPUs as to correctly handle the example above,
and try to maximise the capacity of the CPU we select.

2. prefer_idle and !boosted
These tasks will start at he lowest capacity CPU but go through all
CPUs and try to minimise capacity_orig for the selection of an idle
CPU, which mimics the existing behaviour. When all CPUs are active,
by iterating though all CPUs we're still able to minimise or maximise
spare capacity or utilisation for the considered target or backup CPU.

3. !prefer_idle and boosted and
4. !prefer_idle and !boosted
Given that it is established that non-latency sensitive (!prefer_idle)
tasks should try to minimise capacity (maximise energy efficiency) and
they consider all CPUs when making this decision, an order in which
CPUs are considered for selection is not necessary.

Change-Id: I940399e22eff29453cba0e2ec52a03b17eec12ae
Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>