ANDROID: sched/fair: always start with the biggest CPU
find_best_target iterates though all CPUs and tries to either maximise
or minimise the capacity of the selected CPU based on the combination
of prefer_idle and boosted information it is provided with.

While the order in which CPUs are iterated though is no longer
necessary, it is important to start with the biggest CPU in order to
cater to the needs of one type of task: a prefer_idle and boosted task.
For this task we have to implement a preference towards big reserved
CPUs that will result in low wakeup latency and lower preemption
probability for this type of tasks.

Therefore, this code removes start_cpu and sets the start point always
to the CPU with the highest capacity_orig. Given that for other cases
we already iterate though all CPU to favour the most energy efficient
ones, always starting with the biggest CPU will not impact the result.

Change-Id: I2ac036226c21855e28e8b065f7182f2eadaa0004
Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Reviewed-by: Patrick Bellasi <patrick.bellasi@arm.com>