PM: Introduce an Energy Model management framework
Several subsystems in the kernel (scheduler and/or thermal at the time
of writing) can benefit from knowing about the power consumed by CPUs.
Yet, this information can come from different sources (DT or SCMI for
example), in different formats, hence making it hard to exploit without
a standard API.

This patch attempts to solve this issue by introducing a centralized
Energy Model framework used to interface the data providers with the
client subsystems. This framework standardizes the API to expose power
costs, and to access them from multiple locations.

The current design assumes that all CPUs in a frequency domain share the
same micro-architecture. As such, the energy model data is structured in
a per-frequency-domain fashion. Drivers aware of frequency domains
(typically, but not limited to, CPUFreq drivers) are expected to register
data in the EM framework using the em_register_freq_domain() API. To do
so, the drivers must provide a callback function that will be called by
the EM framework to populate the tables. As of today only the active
power of the CPUs is considered. For each frequency domain, the Energy
Model includes a list of <frequency, power, capacity> tuples for the
capacity states of the domain alongside a cpumask covering the involved
CPUs.

The EM framework also provides an API to re-scale the capacity values
of the model asynchronously, after it has been created. This is required
for architectures where the capacity scale factor of CPUs can change at
run-time. This is the case for Arm/Arm64 for example where the
arch_topology driver recomputes the capacity scale factors of the CPUs
after the maximum frequency of all CPUs has been discovered. Although
complex, the process of creating and re-scaling the Energy Model has to
be kept in two separate steps to fulfill the needs of the different
users. The thermal subsystem doesn't use the capacity values and
shouldn't have dependencies on subsystems providing them. On the other
hand, the scheduler needs the capacity values, and it will benefit from
seeing them up-to-date when applicable.

Because of this need for asynchronous update, the capacity state table
of each frequency domain is protected by RCU, hence guaranteeing a safe
modification of the table and a fast access to readers in latency-sensitive
code paths.

Cc: Peter Zijlstra <peterz@infradead.org>
Cc: "Rafael J. Wysocki" <rjw@rjwysocki.net>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>